# docs/openapi.json
{"swagger":"2.0","info":{"description":"このAPIは個人開発用で試験的に作ったものです。いわゆるシンプルなCRUDができます。","version":"v1","title":"My Simple API","termsOfService":"????","contact":{"name":"kannna5296","url":"xxx","email":"xxx"},"license":{"name":"xxx","url":"xxxx"}},"host":"localhost","basePath":"/","tags":[{"name":"user-controller","description":"User Controller"}],"paths":{"/user/":{"get":{"tags":["user-controller"],"summary":"getUser","operationId":"getUserUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ReadUserResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["user-controller"],"summary":"createUser","operationId":"createUserUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"req","description":"req","required":true,"schema":{"$ref":"#/definitions/CreateUserRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CreateUserResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["user-controller"],"summary":"createUser","operationId":"createUserUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"req","description":"req","required":true,"schema":{"$ref":"#/definitions/UpdateUserRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EmptyResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["user-controller"],"summary":"deleteUser","operationId":"deleteUserUsingDELETE","produces":["*/*"],"parameters":[{"in":"body","name":"req","description":"req","required":true,"schema":{"$ref":"#/definitions/UpdateUserRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EmptyResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"CreateUserRequest":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}},"title":"CreateUserRequest"},"CreateUserResponse":{"type":"object","properties":{"createdAt":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"},"updatedAt":{"type":"string"}},"title":"CreateUserResponse"},"EmptyResponse":{"type":"object","title":"EmptyResponse"},"ReadUserResponse":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/definitions/User"}}},"title":"ReadUserResponse"},"UpdateUserRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"newName":{"type":"string"}},"title":"UpdateUserRequest"},"User":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"password":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"User"}}}